refactor(odata2ts)!: bundle v2 and v4 options
Currently we have 3 options which only apply to V2 and 3 options which only apply to V4; this creates a v2 and v4 option block to hold the respective options

BREAKING CHANGE: v2ResponseResultsWrapping, v2PayloadResultsWrapping, v2ResponseAsV4 are now under prop "v2" and renamed to "responseResultsWrapping", "payloadResultsWrapping", "responseAsV4"; v4BigNumberAsString, enableNativeInOperator, odataVersionV4 are now under prop "v4" and renamed to "bigNumberAsString", "odataVersion" and still "enableNativeInOperator"
